Move `EndianBuf` into its own module
